+++ templates/campaign/donate-modal-window.php
@@ -21,7 +21,7 @@ $modal_class = apply_filters( 'charitable_modal_window_class', 'charitable-modal
 wp_print_scripts( 'lean-modal' );
 wp_enqueue_style( 'lean-modal-css' );
 ?>
-<div id="charitable-donation-form-modal" style="display: none;" class="<?php echo esc_attr( $modal_class ) ?>">
+<div id="charitable-donation-form-modal" style="display: none;" class="<?php echo esc_attr( $modal_class ); ?>">
 	<a class="modal-close"></a>
-	<?php $campaign->get_donation_form()->render() ?>
+	<?php $campaign->get_donation_form()->render(); ?>
 </div>