+++ templates/donation-receipt/not-authorized.php
@@ -7,21 +7,23 @@
  * @author  Studio 164a
  * @package Charitable/Templates/Donation Receipt
  * @since   1.1.2
- * @version 1.1.2
+ * @version 1.6.26
  */
 
 // Exit if accessed directly.
-if ( ! defined( 'ABSPATH' ) ) { exit; }
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
 
 $content = $view_args['content'];
 
 if ( is_user_logged_in() ) : ?>
 	<div class="charitable-notice">
-		<?php _e( 'You do not have access to this donation receipt.', 'charitable' ) ?>
+		<?php _e( 'You do not have access to this donation receipt.', 'charitable' ); ?>
 	</div>
 <?php else : ?>
 	<div class="charitable-notice">
-		<?php _e( 'You must be logged in to access your donation receipt.', 'charitable' ) ?>
+		<?php _e( 'You must be logged in to access your donation receipt.', 'charitable' ); ?>
 	</div>
-	<?php charitable_template( 'shortcodes/login.php', array( 'redirect' => charitable_get_current_url() ) ) ?>
+	<?php echo Charitable_Login_Shortcode::display( array( 'redirect' => charitable_get_current_url() ) ); ?>
 <?php endif ?>